sanity test for dqn
